src/tabulator/server/requirements.txt
requests==2.18.4
opencv_contrib_python==4.5.1.48
six==1.11.0
moviepy==1.0.3
Flask==1.1.2
numpy==1.13.3
Pyrebase==3.0.27
python_magic==0.4.22
